chore: preview for the documentation site structure
The preview displays a single branch (the latest declared one) for each
component and publish it on https://surge.sh as we do for PR preview in the
documentation content repository.
The built site with a single branch per component size is `105.5Mb`.

Other refactorings and improvements
  - enforce 'no robots' on preview
  - bash script: only fetch the documentation content repository when running on
  CI environment
  - antora playbook: use the list array syntax for branches declaration
  Simplify diff management (avoid single line changes that occurred with the
  previous single line array declaration). This is the same syntax used by
  yaml.js so easier diff with generated file.
