+++ .github/workflows/publish-pr-preview.yml
@@ -1,0 +1,51 @@
+name: Publish PR preview
+
+on:
+  pull_request:
+    # To manage 'surge-preview' action teardown, add default event types + closed event type
+    types: [opened, synchronize, reopened, closed]
+    branches:
+      - master
+    paths:
+      - 'resources/*'
+      - 'antora-playbook.yml'
+      - '.github/workflows/generate-documentation.yml'
+jobs:
+  build_preview:
+    runs-on: ubuntu-20.04
+    env:
+      PR_NUMBER: ${{ github.event.pull_request.number }}
+#      GOOGLE_ANALYTICS_KEY: G-7LLM509FEH
+    steps:
+      - uses: actions/checkout@v2
+        if: github.event.action != 'closed'
+      - name: Setup node
+        if: github.event.action != 'closed'
+        uses: actions/setup-node@v1
+        with:
+          node-version: 12.x
+      - uses: actions/cache@v2
+        if: github.event.action != 'closed'
+        with:
+          path: ~/.npm
+          key: ${{ runner.os }}-node-${{ hashFiles('**/package-lock.json') }}
+          restore-keys: |
+            ${{ runner.os }}-node-
+      - name: Compute environment variables
+        run: |
+          echo "BRANCH_NAME=$(echo ${GITHUB_HEAD_REF#refs/heads/})" >> $GITHUB_ENV
+          # the surge-preview action generates https://{{repository.owner}}-{{repository.name}}-{{job.name}}-pr-{{pr.number}}.surge.sh
+          repo_owner_and_name=$(echo "${{github.repository}}" | sed 's/\//-/g')
+          echo PREVIEW_URL=https://$repo_owner_and_name-"${{github.job}}"-pr-$PR_NUMBER.surge.sh >> $GITHUB_ENV
+      - name: Publish preview
+        uses: afc163/surge-preview@v1
+        with:
+          surge_token: ${{ secrets.SURGE_TOKEN_DOC }}
+          github_token: ${{ secrets.GITHUB_TOKEN }}
+          dist: build/site
+          failOnError: true
+          teardown: 'true'
+#        GOOGLE_ANALYTICS_KEY=${{env.GOOGLE_ANALYTICS_KEY}} xxxx
+          build: |
+            ./build-preview.bash --single- --pr "${{ env.PR_NUMBER }}" --site-url "${{ env.PREVIEW_URL }}" --single-branch-per-repo true
+            ls -lh build/site
