+++ antora-playbook.yml
@@ -11,16 +11,30 @@ runtime:
   cache_dir: ./.cache/antora
 
 content:
-  # When adding a new url, add a new entry in generate-doc-content-pr-preview-antora-playbook.js to manage PR preview
+  # When adding a new url, add a new entry in generate-content-for-preview-antora-playbook.js to manage PR preview
+  # The site preview only use the latest declared branch for each source
   sources:
     - url: https://github.com/bonitasoft/bonita-cloud-doc.git
-      branches: [master_asciidoctor]
+      branches:
+        - master_asciidoctor
     - url: https://github.com/bonitasoft/bonita-doc.git
-      branches: [ 7.3_asciidoctor, 7.4_asciidoc, 7.5_new, 7.6_asciidoctor, 7.7_asciidoctor, 7.8_asciidoc, 7.9_asciidoctor, 7.10_asciidoctor, 7.11_asciidoctor, 7.12_asciidoc]
+      branches:
+        - 7.3_asciidoctor
+        - 7.4_asciidoc
+        - 7.5_new
+        - 7.6_asciidoctor
+        - 7.7_asciidoctor
+        - 7.8_asciidoc
+        - 7.9_asciidoctor
+        - 7.10_asciidoctor
+        - 7.11_asciidoctor
+        - 7.12_asciidoc
     - url: https://github.com/bonitasoft/bonita-ici-doc.git
-      branches: [master]
+      branches:
+        - master
     - url: https://github.com/bonitasoft/bonita-continuous-delivery-doc.git
-      branches: [ 3.4_asciidoctor ]
+      branches:
+        - 3.4_asciidoctor
 ui:
   bundle:
     url: ./resources/antora-ui-bundle.zip
