+++ build-preview.bash
@@ -5,28 +5,41 @@
 ########################################################################################################################
 set -euo pipefail
 
-# if arguments contain '--ci false', don't run 'npm ci'
+
+function displayHelpAndExit() {
+
+    exit 0;
+}
+
+
+# Parse arguments
+runOnCI=true
+extraAntoraArgs=
+
 args="$@"
+# if arguments contain '--ci false', don't run 'npm ci'
 if [[ "$args" != *"--ci false"* ]]; then
   echo "Assume script is running on CI environment"
-  echo "Using node $(node --version) and npm $(npm --version)"
-  npm ci
+  extraAntoraArgs="--fetch"
 else
   echo "Assume script is NOT running on CI environment"
+  runOnCI=false
 fi
 
+# if arguments contain '--type local', use html extension in urls
+if [[ "$args" == *"--type local"* ]]; then
+  extraAntoraArgs="${extraAntoraArgs} --html-url-extension-style default"
+fi
+echo "extra Antora Args: ${extraAntoraArgs}"
+
+if [[ $runOnCI == "true" ]]; then
+  echo "Using node $(node --version) and npm $(npm --version)"
+  npm ci
+fi
 # See the node script for the list of arguments
-node scripts/generate-doc-content-pr-preview-antora-playbook.js "$@"
+node scripts/generate-content-for-preview-antora-playbook.js "$@"
 
 echo "Building preview..."
 rm -rf build/
-
-extraAntoraArgs=
-# if arguments contain '--type local', use html extension in urls
-if [[ "$args" == *"--type local"* ]]; then
-  extraAntoraArgs="--html-url-extension-style default"
-fi
-echo "extra Antora Args: ${extraAntoraArgs}"
-./node_modules/.bin/antora --stacktrace --fetch ${extraAntoraArgs} antora-playbook-doc-content-pr-preview.yml
-
+./node_modules/.bin/antora --stacktrace ${extraAntoraArgs} antora-playbook-content-for-preview.yml
 echo "Preview built"
